backend/requirements.txt
fastapi~=0.110.1
uvicorn~=0.29.0
jinja2~=3.1.3
SQLAlchemy[asyncio]~=2.0.23
# fastapi-sqlalchemy
alembic~=1.13.1
openai~=1.55.3
langchain~=0.1.16
langchain-openai~=0.1.3
pypdf~=4.2.0
python-dotenv~=1.0.1
chromadb~=0.4.22
psycopg~=3.1.14
python-multipart==0.0.9
psycopg2==2.9.9
redis~=5.0.1
bcrypt==4.0.1
passlib[bcrypt]~=1.7.4
python-jose~=3.3.0
Unidecode~=1.3.8
beautifulsoup4~=4.12.3
selenium~=4.26.1
webdriver-manager~=4.0.2